requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.31.0
cohere>=4.44
pdfplumber==0.10.3 